requirements_mimir.txt
# 1. create a venv
# 2. pip install -r requirements_mimir.txt

chalice

boto3==1.34.72
langchain==0.1.14
orjson==3.9.15 # needed for langchain > 0.1.7 https://github.com/langchain-ai/langchain/issues/19719
openai==1.14.2
python-dotenv==1.0.0
langchain-community==0.0.30
langchain-openai==0.0.4
langchain-text-splitters
# langchain==0.1.10 #https://github.com/langchain-ai/langchain/issues/18409
# specific
playwright
beautifulsoup4
# chromadb
pypdf
youtube-transcript-api
pinecone-client
langchain-pinecone
